rag/pyproject.toml
[project]
name = "bitrix-rag"
version = "0.0.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.1",
  "httpx>=0.27.0",
  "qdrant-client>=1.9.0",
  "markdown-it-py>=3.0.0",
  "rank-bm25>=0.2.2",
  "python-dotenv>=1.0.1",
]

[project.scripts]
bitrix-rag = "bitrix_rag.cli:main"

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

